Modified standalone examples
