+++ examples/kokkos-tutorials/standalone/subview.py
@@ -6,8 +6,7 @@ from parse_args import parse_args
 
 
 @pk.workunit
-def yAx(j: int, acc: pk.Acc[float], M: int,
-        y: pk.View1D[pk.double], x: pk.View1D[pk.double], A: pk.View2D[pk.double]):
+def yAx(j, acc, M, y, x, A):
     temp2: float = 0
     A_row_j = A[j,:];
     for i in range(M):
